server/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "productive-doomscrolling"
version = "0.1.0"
description = "Turn long-form videos into vertical clips for productive doomscrolling."
authors = [
    { name = "Learning Engineer" }
]
readme = "../README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.2",
    "uvicorn[standard]>=0.27.1",
    "python-multipart>=0.0.9",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.1",
    "groq>=0.4.1",
    "google-genai>=0.1.0",
    "jinja2>=3.1.3",
    "yt-dlp>=2023.12.30",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.2.1",
    "httpx>=0.26.0"
]
